ch4: Rename the VCI attributes to VCI resources
The design of the VCI pool infrastructure (pmodels#3356)
classifies the TX and RX operations as resources of the VCI. This
commit renames the variables and functions related to TX and RX
attributes to reflect the design.

Additionally, the commit also moves the VCI definitions to a
separate file so that all VCI-related code will house under the
`ch4_vci*` prefix.

Signed-off-by: <To be reviewed>