ch4/progress: Define progress function that waits for a request
`MPID_Progress_wait_req` waits for the given request to
complete before returning.

This function requires that the VCI on which an operation that
hasn't completed immediately is stored in the request of that
operation.

TODO: implement the same function in CH3 so the MPIR layer
can use it without build failures for CH3

Signed-off-by: <To be reviewed>