ch4: Rename the VCI attributes to VCI resources
The design of the VCI pool infrastructure (pmodels#3356)
classifies the TX and RX operations as resources of the VCI. This
commit renames the variables and functions related to TX and RX
attributes to reflect the design.

Additionally, this commit also typedefs the VCI resources into an
enum. The `vci_get_resource_info` function now uses this enum.

Signed-off-by: <To be reviewed>