ch4/progress: Modularize progress with thread-safe functions
In terms of code design, this commit modularizes
MPIDI_Progress_test into progress of hooks, and progress of the
VCI. Each of these modularized functions are thread-safe.

In terms of performance, this commit removes an extra lock/unlock
pair on the VCI's lock when a WorkQ is involved. Previously,
progress on the WorkQ was made through a thread-safe function
after which the VCI lock was taken again to make progress on the
VCI. This commit increases the granularity of the critical
section by taking the lock once to progress the VCI and its
WorkQ(s).