ch4: Move per-thread Request cache to per-VCI
Creation of requests need not be safe since the VCI lock is always
held when it is time to create a request.

Freeing of requests during progress also means that the right VCI
lock is held at that time. All the requests of operations
progressed on VCI i need to be returned to VCI i. During the
progress of VCI i, the lock of VCI i is held.

Freeing of requests outside of progress (for example, during Waitall)
requires the lock of the VCI to be taken since there is no guarantee
of the lock being held.