ch4: Use VCI fields directly instead of using them through the window
As we want to move towards using mulitple VCIs per window, we
cannot rely on using the objects belonging to each window since
those are merely pointers to VCI's objects.

To use a VCI's objects direclty, we need to pass down VCI
information to the netmod after hashing it once to prevent
incurring the costs of hashing multiple times.