ch4: Make VCI pool a statically allocated array
The maximum number of VCIs may not be the same as the number of
VCIs actually set up. The number of VCIs set up will depend on
the number of VNIs/VSIs. This idea is the same as the one used
in the netmods. Both netmods statically allocate memory for the
VNI pool but actually use only as many VNIs as specified by the
user through the CVAR.

Shifting to this static allocation approach for two reasons:

1) Removes pointer dereference whenever accessing a VCI
2) For functions (such as an MPI_Testall) where we need to
temporarily store per-VCI information, we don't need to malloc
an array every time.